Ryan/feature/apply settings (#94)
* Main menu file handling (#93)

* Empty settings screen added. TODO:- Content to be added

* Toggle Button added. TODO:- Integrate it in settings screen

* Toggle Animation working

* Event handling completed. TOFIX: - Restrict any event while animating. TODO:- Handle navigation

* Back Navigation Handled. TODO: - File handling

* Freed all ToonToggleButtons in settings destructor

* Added toggle parameter in settingsdatastructure

* Data read from file. TODO:- Writing correct data to the file

* Settings data saved in file.

* gitignore updated

* Toggle value name parameter added

---------


* Added: Invert Camera setting

* Apply Setting: Shadow Quality

---------

+++ CSC8503/ToonGameWorld.h
@@ -64,7 +64,7 @@ namespace NCL
 			}
 
 			int GetMainCameraCount() {
-				return mainCameras.size();
+				return (int)mainCameras.size();
 			}
 
 			Camera* GetMinimapCamera() const { return minimapCamera; }
